Refactor UI Notification System and Introduce ASCII Art Package
- Moved notification utilities from cmd/ui/notify to pkg/ui/notify for better organization and modularity.
- Updated all relevant imports across the codebase to reflect the new package structure.
- Introduced a new pkg/ui/asciiart package to handle ASCII art rendering, including a KSail logo.
- Added tests for the new ASCII art functionality to ensure output consistency and correctness.
- Enhanced documentation for both pkg/ui/notify and pkg/ui/asciiart to clarify usage and features.
- Updated error handling in the notify package to ensure robust messaging in case of output failures.
