+++ cmd/root.go
@@ -5,9 +5,9 @@ import (
 
 	cluster "github.com/devantler-tech/ksail-go/cmd/cluster"
 	"github.com/devantler-tech/ksail-go/cmd/internal/cmdhelpers"
-	"github.com/devantler-tech/ksail-go/cmd/ui/asciiart"
 	"github.com/devantler-tech/ksail-go/cmd/workload"
 	configmanager "github.com/devantler-tech/ksail-go/pkg/config-manager/ksail"
+	"github.com/devantler-tech/ksail-go/pkg/ui/asciiart"
 	"github.com/spf13/cobra"
 )
