+++ pkg/ui/README.md
@@ -1,4 +1,4 @@
-# cmd/ui
+# pkg/ui
 
 This directory contains user interface utilities for KSail's CLI.
 
@@ -15,8 +15,8 @@ Provides UI components and utilities for enhancing the command-line user experie
 
 ## Packages
 
-- **[cmd/ui/asciiart/](./asciiart/README.md)** - ASCII art and visual elements for CLI branding
-- **[cmd/ui/notify/](./notify/README.md)** - User notification and messaging utilities with colored output
+- **[pkg/ui/asciiart/](./asciiart/README.md)** - ASCII art and visual elements for CLI branding
+- **[pkg/ui/notify/](./notify/README.md)** - User notification and messaging utilities with colored output
 
 ## Usage
 
@@ -24,8 +24,8 @@ These packages are used internally by CLI commands to provide a consistent and u
 
 ```go
 import (
-    "github.com/devantler-tech/ksail-go/cmd/ui/notify"
-    "github.com/devantler-tech/ksail-go/cmd/ui/asciiart"
+    "github.com/devantler-tech/ksail-go/pkg/ui/notify"
+    "github.com/devantler-tech/ksail-go/pkg/ui/asciiart"
 )
 
 // Provide user feedback
