+++ pkg/ui/asciiart/ksail_logo.go
