+++ pkg/ui/asciiart/ksail_logo.txt
